Automatically wrap non-vararg specs in s/spec
This allows the seamless situation of defn-spec args which use `s/cat`
or `s/*` or other regex-like specs. Without the implicit `s/spec`
wrapper, the meaning of those specs will change to apply to the argument
list itself, which is almost certainly never what's intended *except*
for the case of the varargs. So, you can finish off your fn with
`& args (s/* any?)` and defn-spec will allow it through.